Update tests to work without first event signal
